requirements.txt
aws-msk-iam-sasl-signer-python==1.0.2
boto3==1.34.23
confluent-kafka==2.8.0
dotenv==0.9.9
great-expectations==1.3.3
psycopg2==2.9.10
